requirements.txt
streamlit
fastembed
faiss-cpu
groq
PyPDF2
numpy 
python-dotenv